✨ refacto auth files + handle db verif for user
